requirements.txt
Flask
gunicorn
line-bot-sdk
googletrans==4.0.0rc1
json
